requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb
pypdf
python-dotenv
sentence-transformers